requirements.txt
cloudscraper~=1.2.71
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
pymongo~=4.6.3